move get_limited_unpacker to helpers
also: move some constants to borg.constants
(cherry picked from commit 89f3cab6cd44903e096fe0237e9b0cc1eeac1e87)
